changed mentality, hold off on automation
The vision is for a user to be able to basically point at a file and say
"make this into this format". That raises the complexity quite a bit,
though. So since I made the `{from,to}_format` cli options mandatory
now, let's ride with that and trust (at least in the proof-of-concept)
that the input file/dir/etc is what the user says it is. Can add
"simple" checks when this is a bit more mature than a ball of clay, and
later on consider more automation/handholding. But let's roll with this
for now.