fixed the h5 bug
I.e. the bug that stopped my generated h5 files to be imported by
scanpy. The issue was this: if there is no genome information supplied
in the features file, I set the h5 dataset for "feature_type" to an
array of "unknown", since scanpy complained when the column had empty
values. I construct the array based on the shape of the data matrix.
However, ofc I had used the columns (nbr of cells) for it instead of the
rows (nbr of features) (`matrix.shape[1]` -> `matrix.shape[0]`). Seurat
apparently doesn't read in that data, hence why that worked, but scanpy
does, which is why that didn't work. But now it does work!