autoformatting, tests for scanpy+seurat read-in etc
broke out the extraction of features in `mtx_to_h5` to its own function.
The if cases there are a bit repetitive, could probably be re-written.
When this is in its own function it is easier to write tests for it in
isolation and ensure the same functionality before, during and after
re-write(s). Plus I can now hide it with folding..

Plus, added (very basic) tests for reading in generated files with
scanpy and Seurat, this is what in the real world will most happen with
these files so it makes a lot of sense to include that in the tests.