testing one approach for *_to_seurat_object
the rpy2 approach. It seems like it is more hassle than it's worth.
Might be better to just check if there is an R installation (in the
Docker container there is one, and expose a flag for the CLI for a user
to supply their own installation) and just call R code. Too much
in-depth to grok rpy2 and I don't see any really convincing arguments
for investing that time